requirements-ui.txt
flask>=3.0
werkzeug>=3.0
python-docx>=1.0
